requirements.txt
pymupdf>=1.23.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
pdfminer.six>=20221105
deep-translator>=1.11.0
